fix uninstaller

+++ uninstall.sh
@@ -2,7 +2,25 @@
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 SERVICE_NAME=$(basename $SCRIPT_DIR)
 
-rm /service/$SERVICE_NAME
-kill $(pgrep -f 'supervise dbus-growatt-shinex.py')
+
 chmod a-x $SCRIPT_DIR/service/run
-./restart.sh
+
+if [ -f /service/$SERVICE_NAME ]
+then
+        rm /service/$SERVICE_NAME
+fi
+
+
+pids=$(pgrep -f $SERVICE_NAME)
+if [ ! -z "$pids" ]
+then
+  for pid in $pids
+  do
+    kill $pid
+  done
+fi
+
+if [ -f '/data/rc.local' ]
+then
+  sed -i '/.*'$SERVICE_NAME'.*/d' /data/rc.local
+fi
